Improving Historian http-related logs
