+++ server/historian/packages/historian-base/package.json
@@ -25,14 +25,14 @@
   },
   "dependencies": {
     "@fluidframework/common-utils": "^0.31.0",
-    "@fluidframework/gitresources": "^0.1033.0-40116",
+    "@fluidframework/gitresources": "^0.1034.0",
     "@fluidframework/protocol-definitions": "^0.1026.0",
-    "@fluidframework/server-services": "^0.1033.0-40116",
-    "@fluidframework/server-services-client": "^0.1033.0-40116",
-    "@fluidframework/server-services-core": "^0.1033.0-40116",
-    "@fluidframework/server-services-shared": "^0.1033.0-40116",
-    "@fluidframework/server-services-telemetry": "^0.1033.0-40116",
-    "@fluidframework/server-services-utils": "^0.1033.0-40116",
+    "@fluidframework/server-services": "^0.1034.0",
+    "@fluidframework/server-services-client": "^0.1034.0",
+    "@fluidframework/server-services-core": "^0.1034.0",
+    "@fluidframework/server-services-shared": "^0.1034.0",
+    "@fluidframework/server-services-telemetry": "^0.1034.0",
+    "@fluidframework/server-services-utils": "^0.1034.0",
     "body-parser": "^1.17.2",
     "compression": "^1.7.3",
     "cors": "^2.8.5",
@@ -50,7 +50,7 @@
   "devDependencies": {
     "@fluidframework/build-common": "^0.23.0",
     "@fluidframework/eslint-config-fluid": "^0.23.0",
-    "@fluidframework/server-test-utils": "^0.1033.0-40116",
+    "@fluidframework/server-test-utils": "^0.1034.0",
     "@types/compression": "0.0.36",
     "@types/cors": "^2.8.4",
     "@types/debug": "^4.1.5",
