+++ server/historian/packages/historian-base/src/utils.ts
@@ -51,6 +51,43 @@ export function getTenantIdFromRequest(params: Params) {
     return "-";
 }
 
+export function getDocumentIdFromRequest(tenantId: string, authorization: string) {
+    if (!tenantId || tenantId === "-" || !authorization) {
+        return "-";
+    }
+
+    const token = parseToken(tenantId, authorization);
+    const decoded = jwt.decode(token) as ITokenClaims;
+
+    if (!decoded) {
+        return "-";
+    }
+
+    return decoded.documentId;
+}
+
+export function parseToken(tenantId: string, authorization: string): string {
+    let token: string;
+    if (authorization) {
+        // eslint-disable-next-line @typescript-eslint/prefer-regexp-exec
+        const base64TokenMatch = authorization.match(/Basic (.+)/);
+        if (!base64TokenMatch) {
+            throw new NetworkError(403, "Malformed authorization token");
+        }
+        const encoded = Buffer.from(base64TokenMatch[1], "base64").toString();
+
+        // eslint-disable-next-line @typescript-eslint/prefer-regexp-exec
+        const tokenMatch = encoded.match(/(.+):(.+)/);
+        if (!tokenMatch || tenantId !== tokenMatch[1]) {
+            throw new NetworkError(403, "Malformed authorization token");
+        }
+
+        token = tokenMatch[2];
+    }
+
+    return token;
+}
+
 /**
  * Pass into `.catch()` block of a RestWrapper call to output a more standardized network error.
  * @param url request url to be output in error log
@@ -59,7 +96,8 @@ export function getTenantIdFromRequest(params: Params) {
  */
 export function getRequestErrorTranslator(
     url: string,
-    method: string): (error: any) => never {
+    method: string,
+    lumberProperties?: Map<string, any> | Record<string, any>): (error: any) => never {
     const standardLogErrorMessage = `[${method}] Request to [${url}] failed`;
     const requestErrorTranslator = (error: any): never => {
         // BasicRestWrapper only throws `AxiosError.response.status` when available.
@@ -67,7 +105,7 @@ export function getRequestErrorTranslator(
         if (typeof error === "number" || !Number.isNaN(Number.parseInt(error, 10)))  {
             const errorCode = typeof error === "number" ? error : Number.parseInt(error, 10);
             winston.error(`${standardLogErrorMessage}: ${errorCode}`);
-            Lumberjack.error(`${standardLogErrorMessage}: ${errorCode}`);
+            Lumberjack.error(`${standardLogErrorMessage}: ${errorCode}`, lumberProperties);
             throw new NetworkError(
                 errorCode,
                 "Internal Service Request Failed",
@@ -75,7 +113,7 @@ export function getRequestErrorTranslator(
         }
         // Treat anything else as an internal error, but log for debugging purposes
         winston.error(`${standardLogErrorMessage}: ${safeStringify(error)}`);
-        Lumberjack.error(standardLogErrorMessage, undefined, error);
+        Lumberjack.error(standardLogErrorMessage, lumberProperties, error);
         throw new NetworkError(500, "Internal Server Error");
     };
     return requestErrorTranslator;
